feat: 옵션 기능을 알림 페이지 작업에서 선행 작업된 코드에 업데이트
